src/requirements.txt
pandas
numpy
scikit-learn  # Si usas ML
openai==0.28
python-dotenv
